+++ src/precompile.jl
@@ -1,4 +1,4 @@
 function _precompile_()
     ccall(:jl_generating_output, Cint, ()) == 1 || return nothing
-    precompile(FastTransforms.pochhammer, (Float64, Int64,))
+    precompile(FastTransforms.pochhammer, (Float64, Int,))
 end
